requirements.txt
google-cloud-vision==3.5.0
google-auth==2.26.2
openai==1.8.0
pydantic==2.5.3
python-dotenv==1.0.0
streamlit==1.30.0
requests==2.31.0
streamlit-extras==0.3.6